Added cli flags to choose the type of reflection options to use during call graph construction.
Changed UriPrefixContextSelector to extend the DefaultContextSelector and use its getCalleeTarget and getRelavantParameters when we aren't interested in string and uri contexts.  This should hopefully fix the errors we used to get when using reflection options.

Refactored the WDexClassLoaderImpl to @Override init instead of loadAllClasses.  Should make it so we don't need the patch to change loadAllClasses from private to protected.

Changed DexIRFactory to extend the DefaultIRFactory.  This change, I believe, makes it so we don't need the DexIContextInterpreter anymore.

The analysis was also using two different analysiscache's.  One from the default new AnalysisCache() constructor and one from new AnalysisCache((IRFactory<IMethod>)new DexIRFactory()).  The latter was located in DexIContextInterpreter.  These two cache's were being used when creating the SSAPropagationCallGraphBuilder object.  I believe there should have been only one cache being used instead of the two that was implemented before.